Transfer CharacterGetter class into Character file
Made this transfer for more coherence
